+++ core/src/main/java/com/ensody/reactivestate/StateCore.kt
@@ -1,0 +1,20 @@
+package com.ensody.reactivestate
+
+import kotlinx.coroutines.CoroutineScope
+
+/**
+ * A state container whose lifetime is bound to a given [CoroutineScope].
+ *
+ * This is useful in UI-based code where you want to keep state and business logic separate from the
+ * UI and the UI framework.
+ *
+ * Since the lifetime is bound to [scope] you can depend on automatic cleanups e.g. when using
+ * [autoRun] or [workQueue].
+ */
+abstract class State(val scope: CoroutineScope) : Disposable {
+    val disposable = disposeOnCompletionOf(scope)
+
+    override fun dispose() {
+        disposable.dispose()
+    }
+}