+++ reactivestate/src/main/java/com/ensody/reactivestate/DerivedLiveData.kt
@@ -21,3 +21,6 @@ fun <T> CoroutineScope.derived(observer: AutoRunCallback<T>) =
 
 fun <T> ViewModel.derived(observer: AutoRunCallback<T>) =
     viewModelScope.derived(observer)
+
+fun <T> State.derived(observer: AutoRunCallback<T>) =
+    scope.derived(observer)